Backend working (not sift)

+++ backend/node.hpp
@@ -33,7 +33,7 @@ struct Container {
 
 struct Node {
 
-    double cost;
+    int cost;
     double heuristic;
     double total;
 	int numToLoad = 0;
